feat: Implement Home Screen with Dashboard Components
- Added HomeScreen widget with a structured layout for the dashboard.
- Integrated various widgets including DailySalesCard, StatsGrid, DailySalesChart, TopSellingProducts, SalesByCategory, and RecentTransactions.
- Created DailySalesChart to visualize daily sales over the last 7 days.
- Implemented LoadingErrorWidget for error handling during data loading.
- Developed RecentTransactions widget to display the latest transactions with navigation to details.
- Added SalesByCategory widget to show sales categorized by menu types.
- Introduced StatsGrid to display key statistics such as Total Sold, Monthly Income, Total Stock, and Low Stock.
- Created individual stat widgets: TotalSoldStat, MonthlyIncomeStat, TotalStockStat, and LowStockStat for modularity.
- Developed TopSellingProducts widget to showcase the top-selling items of the current month.
- Utilized Bloc pattern for state management across the dashboard components.
