+++ lib/screens/home_screen/home_screen.dart
@@ -1,0 +1,121 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:the_djenggot/bloc/menu/menu_bloc.dart';
+import 'package:the_djenggot/bloc/menu/menu_event.dart';
+import 'package:the_djenggot/bloc/stock/stock_bloc.dart';
+import 'package:the_djenggot/bloc/transaction/transaction_bloc.dart';
+import 'package:the_djenggot/bloc/transaction/transaction_event.dart';
+import 'package:the_djenggot/bloc/transaction/transaction_state.dart';
+import 'package:the_djenggot/bloc/type/menu_type/menu_type_bloc.dart';
+import 'package:the_djenggot/bloc/type/menu_type/menu_type_event.dart';
+import 'package:the_djenggot/bloc/type/stock_type/stock_type_bloc.dart';
+import 'package:the_djenggot/bloc/type/stock_type/stock_type_event.dart';
+import 'package:the_djenggot/utils/theme/app_theme.dart';
+import 'package:the_djenggot/widgets/transaction/daily_sales_card.dart';
+
+// Import modular components
+import 'widgets/stats_grid.dart';
+import 'widgets/daily_sales_chart.dart';
+import 'widgets/top_selling_products.dart';
+import 'widgets/sales_by_category.dart';
+import 'widgets/recent_transactions.dart';
+
+class HomeScreen extends StatefulWidget {
+  final PageController pageController;
+
+  const HomeScreen({super.key, required this.pageController});
+
+  @override
+  State<HomeScreen> createState() => _HomeScreenState();
+}
+
+class _HomeScreenState extends State<HomeScreen> {
+  DateTime selectedDate = DateTime.now();
+
+  @override
+  void initState() {
+    super.initState();
+    refresh();
+  }
+
+  void refresh() {
+    context.read<MenuBloc>().add(LoadMenu());
+    context.read<StockBloc>().add(LoadStock());
+    context.read<MenuTypeBloc>().add(LoadMenuTypes());
+    context.read<StockTypeBloc>().add(LoadStockTypes());
+    context.read<TransactionBloc>().add(LoadTransactions());
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text(
+          "Dashboard",
+          style: AppTheme.appBarTitle,
+        ),
+      ),
+      body: BlocListener<TransactionBloc, TransactionState>(
+        listener: (context, state) {},
+        child: RefreshIndicator(
+          onRefresh: () async {
+            refresh();
+          },
+          child: ListView(
+            padding: const EdgeInsets.all(16),
+            children: [
+              // Daily sales card
+              BlocBuilder<TransactionBloc, TransactionState>(
+                builder: (context, state) {
+                  if (state is TransactionLoaded) {
+                    return DailySalesCard(
+                      transactions: state.transactions,
+                      onDateChanged: (date) {
+                        setState(() {
+                          selectedDate = date;
+                        });
+                      },
+                      selectedDate: selectedDate,
+                    );
+                  } else if (state is TransactionLoading) {
+                    return const Center(child: CircularProgressIndicator());
+                  } else {
+                    return const Center(child: Text("Error loading transactions"));
+                  }
+                },
+              ),
+
+              const SizedBox(height: 32),
+
+              // Stats grid with 4 cards
+              const StatsGrid(),
+
+              const SizedBox(height: 32),
+
+              // Daily sales chart
+              const DailySalesChart(),
+
+              const SizedBox(height: 32),
+
+              // Top selling products
+              const TopSellingProducts(),
+
+              const SizedBox(height: 32),
+
+              // Sales by category
+              const SalesByCategory(),
+
+              const SizedBox(height: 32),
+
+              // Recent transactions
+              RecentTransactions(
+                pageController: widget.pageController,
+                onRefresh: refresh,
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
