+++ lib/screens/home_screen/widgets/loading_error_widget.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+
+class LoadingErrorWidget extends StatelessWidget {
+  const LoadingErrorWidget({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const SizedBox(
+      height: 30,
+      child: Center(
+        child: SizedBox(
+          height: 15,
+          width: 15,
+          child: CircularProgressIndicator(
+            strokeWidth: 2,
+          ),
+        ),
+      ),
+    );
+  }
+}
